DEX-378 Errors formatting fixes (#94)
MatcherApiRoute: handling client's requests with a wrong JSON ;

MatcherError:
* Introduced error.InvalidJson;
* Amount and Price contain BigDecimal now to properly format their values;

MatchingRule.tickSize now is BigDecimal;

History: Double is replaced by BigDecimal to properly format corner cases;

Unit tests: MatcherApiRouteSpec - added cases with a wrong Order's JSON;

Integration tests: OrderRestrictionsTestSuite - "high amount" now contains a huge amount that improperly formatted before;

+++ dex/src/main/scala/com/wavesplatform/dex/history/DBRecords.scala
@@ -14,22 +14,22 @@ object DBRecords {
                          priceAssetId: String,
                          feeAssetId: String,
                          side: Byte,
-                         price: Double,
-                         amount: Double,
+                         price: BigDecimal,
+                         amount: BigDecimal,
                          timestamp: LocalDateTime,
                          expiration: LocalDateTime,
-                         fee: Double,
+                         fee: BigDecimal,
                          created: LocalDateTime)
       extends Record
 
   case class EventRecord(orderId: String,
                          eventType: Byte,
                          timestamp: LocalDateTime,
-                         price: Double,
-                         filled: Double,
-                         totalFilled: Double,
-                         feeFilled: Double,
-                         feeTotalFilled: Double,
+                         price: BigDecimal,
+                         filled: BigDecimal,
+                         totalFilled: BigDecimal,
+                         feeFilled: BigDecimal,
+                         feeTotalFilled: BigDecimal,
                          status: Byte)
       extends Record
 }
